+++ karma.conf.js
@@ -9,16 +9,19 @@ module.exports = function (config) {
       require('karma-jasmine'),
       require('karma-chrome-launcher'),
       require('karma-jasmine-html-reporter'),
-      require('karma-coverage'),
       require('@angular-devkit/build-angular/plugins/karma')
     ],
     client: {
+      jasmine: {
+        // you can add configuration options for Jasmine here
+        // the possible options are listed at https://jasmine.github.io/api/edge/Configuration.html
+        // for example, you can disable the random execution with `random: false`
+        // or set a specific seed with `seed: 4321`
+      },
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
-    coverageIstanbulReporter: {
-      dir: require('path').join(__dirname, './coverage/ngx-progressbar'),
-      reports: ['html', 'lcovonly', 'text-summary'],
-      fixWebpackSourcePaths: true
+    jasmineHtmlReporter: {
+      suppressAll: true // removes the duplicated traces
     },
     reporters: ['progress', 'kjhtml'],
     port: 9876,
@@ -36,9 +39,9 @@ module.exports = function (config) {
     restartOnFileChange: true,
     capabilities: {
       chromeOptions: {
-        args: [ "--headless" ]
+        args: ["--headless"]
       },
-      'browserName': 'chrome'
-    }
+      browserName: 'chrome'
+    },
   });
 };