+++ projects/ngx-progressbar-demo/src/app/custom/custom.component.ts
@@ -1,5 +1,11 @@
 import { ChangeDetectionStrategy, Component, OnDestroy, OnInit } from '@angular/core';
+import { CommonModule } from '@angular/common';
 import { animate, state, style, transition, trigger } from '@angular/animations';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatCardModule } from '@angular/material/card';
 import { NgProgress, NgProgressRef } from 'ngx-progressbar';
 
 @Component({
@@ -15,7 +21,9 @@ import { NgProgress, NgProgressRef } from 'ngx-progressbar';
       ])
     ])
   ],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [CommonModule, MatCardModule, MatProgressBarModule, MatButtonModule, MatIconModule, MatProgressSpinnerModule]
 })
 export class CustomComponent implements OnInit, OnDestroy {