+++ projects/ngx-progressbar-demo/src/app/home/lab/lab.component.ts
@@ -1,12 +1,15 @@
 import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 import { NgProgressConfig } from 'ngx-progressbar';
-// import { NgProgressConfig } from '../../../../../ngx-progressbar/src/public-api';
 
 @Component({
   selector: 'app-lab',
   templateUrl: './lab.component.html',
   styleUrls: ['./lab.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [FormsModule, CommonModule]
 })
 export class LabComponent {