+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 11.1.0
+
+- feat: Migrate to Angular standalone components in [1a6f780](https://github.com/MurhafSousli/ngx-progressbar/pull/360/commits/1a6f780c2796b5762c47bd25f38079f613783dfe).
+- refactor: Update rxjs imports in [6f426b1](https://github.com/MurhafSousli/ngx-progressbar/pull/360/commits/6f426b165187a5791da13f823903eff9fa91f409).
+
 ## 11.0.0
 
 - Update to Angular 16 in [5f05486](https://github.com/MurhafSousli/ngx-progressbar/pull/357/commits/5f054867396415c5490592bf6a3d13d68d1aa62e).